+++ include/i2c_bus.h
@@ -1,0 +1,6 @@
+#ifndef __I2C_BUS_H__
+#define __I2C_BUS_H__
+
+void i2c_bus_init(void);
+
+#endif