+++ kernel/i2c_bus.c
@@ -1,0 +1,19 @@
+#include "driver/i2c.h"
+#include "config.h"
+#include "i2c_bus.h"
+
+#define I2C_MASTER_FREQ_HZ 100000
+
+void i2c_bus_init(void)
+{
+    i2c_config_t conf = {
+        .mode = I2C_MODE_MASTER,
+        .sda_io_num = (gpio_num_t)CONFIG_I2C_SDA_PIN,
+        .scl_io_num = (gpio_num_t)CONFIG_I2C_SCL_PIN,
+        .sda_pullup_en = GPIO_PULLUP_ENABLE,
+        .scl_pullup_en = GPIO_PULLUP_ENABLE,
+        .master.clk_speed = I2C_MASTER_FREQ_HZ,
+    };
+    ESP_ERROR_CHECK(i2c_param_config(I2C_NUM_1, &conf));
+    ESP_ERROR_CHECK(i2c_driver_install(I2C_NUM_1, I2C_MODE_MASTER, 0, 0, 0));
+}